Fix derp (only listen for Nucleus events if Nucleus is present). getHighestGroup - Only return group identifier if is not excluded via config (otherwise return empty string).
