+++ src/main/java/br/net/fabiozumbi12/rankupper/PermsAPI.java
@@ -16,30 +16,19 @@ public class PermsAPI {
 		this.permissionService = game.getServiceManager().getRegistration(PermissionService.class).get().getProvider();
 	}
 	
-	public List<String> getGroups(User player){
-		List<String> gps = new ArrayList<String>();
-		for (Subject sub:player.getParents()){
-			if (sub.getContainingCollection().equals(getGroups()) && (sub.getIdentifier() != null)){
-				gps.add(sub.getIdentifier());
-			}
-		}
-		return gps;
-	}
-	
 	public String getHighestGroup(User player){
-		String best = "";
-		int beb = -1;
-		for (Subject sub:player.getParents()){
-			if (sub.getContainingCollection().equals(getGroups()) && (sub.getIdentifier() != null)){
+		for (Subject sub:player.getParents()) {
+			if (sub.getContainingCollection().equals(getGroups()) && (sub.getIdentifier() != null)) {
 				RULogger.debug(sub.getIdentifier() + " - " + sub.getParents().size());
-				if(beb < sub.getParents().size()){
-					beb = sub.getParents().size();
-					best = sub.getIdentifier();
+				for (String excl : RankUpper.cfgs.getStringList("excluded-groups")) {
+					if (excl.equalsIgnoreCase(sub.getIdentifier())) {
+						break;
+					}
 				}
+				return sub.getIdentifier();
 			}
 		}
-		//return gps;
-		return best;
+		return "";
 	}
 	
 	public SubjectCollection getGroups(){
