+++ src/main/java/br/net/fabiozumbi12/rankupper/RUAFK.java
@@ -1,7 +1,8 @@
 package br.net.fabiozumbi12.rankupper;
 
+import io.github.nucleuspowered.nucleus.api.events.NucleusAFKEvent;
 import org.spongepowered.api.entity.living.player.Player;
-
+import org.spongepowered.api.event.Listener;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -11,6 +12,7 @@ import java.util.List;
 public class RUAFK
 {
     private List<Player> afkPlayers;
+    RankUpper plugin;
 
     RUAFK() {
         this.afkPlayers = new ArrayList<>();
@@ -24,6 +26,26 @@ public class RUAFK
     }
     public boolean isPlayer(Player p){
         return this.afkPlayers.contains(p);
+    }
+
+    @Listener
+    public void PlayerGoingAFK(NucleusAFKEvent.GoingAFK e) {
+        if(RankUpper.cfgs.getBool("afk-support")) {
+            Player p = e.getTargetEntity();
+            if(!plugin.getRUAFK().isPlayer(p)){
+                plugin.getRUAFK().addPlayer(p);
+            }
+        }
+    }
+
+    @Listener
+    public void PlayerReturningFromAFK(NucleusAFKEvent.ReturningFromAFK e) {
+        if(RankUpper.cfgs.getBool("afk-support")) {
+            Player p = e.getTargetEntity();
+            if(plugin.getRUAFK().isPlayer(p)){
+                plugin.getRUAFK().removePlayer(p);
+            }
+        }
     }
 
 }
